trading sb back end completed once again need to check all the functionalities

+++ trading-sb/src/main/java/com/akhil/trading/model/OrderItem.java
@@ -16,7 +16,7 @@ public class OrderItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private double quantity;
+    private BigDecimal quantity;
 
     private String coinId;
