+++ trading-sb/src/main/java/com/akhil/trading/service/AssetService.java
@@ -3,10 +3,11 @@ package com.akhil.trading.service;
 import com.akhil.trading.model.Asset;
 import com.akhil.trading.model.Coin;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AssetService {
-    void createAsset(Long userId, Coin coin, double quantity);
+    void createAsset(Long userId, Coin coin, BigDecimal quantity);
 
     Asset getAssetById(Long assetId);
 
@@ -14,7 +15,7 @@ public interface AssetService {
 
     List<Asset> getUsersAssets(Long userId);
 
-    Asset updateAsset(Long assetId,double quantity);
+    Asset updateAsset(Long assetId,BigDecimal quantity);
 
     Asset findAssetByUserIdAndCoinId(Long userId,String coinId);
