+++ trading-sb/src/main/java/com/akhil/trading/service/OrderService.java
@@ -5,12 +5,13 @@ import com.akhil.trading.model.Coin;
 import com.akhil.trading.model.Order;
 import com.akhil.trading.model.OrderItem;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OrderService{
     Order createOrder(Long  userId, OrderItem orderItem, OrderType orderType);
     Order getOrderById(Long orderId);
     List<Order> getAllOrderOfUser(Long userId,String orderType,String assetSymbol);
-    Order processOrder(Coin coin, double quantity,OrderType orderType,Long userId);
+    Order processOrder(Coin coin, BigDecimal quantity, OrderType orderType, Long userId);
 
 }
